fix(build): resolve all remaining compiler errors (ARM GCC build fully green)
Completes the third and final pass over the remaining ~46 error lines
(563 error lines / 474 unique at the start of this pass -> 0 errors).
All targets build 100% with -j4 and -j1.

## Highlights
- CMake: add missing include dirs for mcal/ecual/services/boot/os/platform
  (platform s32k312, mcal can/spi/fls/lin/uart, third_party mbedtls/aes_modes/hash,
  include/autosar, config/input/*)
- New stub headers: FreeRTOSConfig/projdefs/portmacro, Reg_Macros, Lockstep,
  PduR_DoIP/PduR_LinTp, Eth_GeneralTypes, SchM_{DoIP,Fee,SecOC,Uart},
  *_MemMap stubs, mbedtls/* (bignum/ecp/ecdsa/ecdh/aes/gcm/sha256/md/hkdf/
  ctr_drbg/entropy/x509/x509_crt/oid/asn1/ssl/pk/platform/debug),
  CryptoStack_Types, Dma
- Remove shadowing autosar stubs (SomeIp/SomeIpTp/SomeIpXf/LinTp/RamSafety)
  so the real module headers are used
- Shared types: ComStack_Types (NetworkHandleType, TP_*, BUFREQ_E_OVFL,
  RetryInfoType), Crypto_Types cleanup, Eth_GeneralTypes
- Complete files truncated by earlier commits: NvM.c, StbM.c, SecOc.c,
  SomeIpTp.c, SomeIpXf.c (restored missing function tails)
- ARM asm made host/target portable (dsb sy / daifset on aarch64)
- Module fixes: Mqtt (TcpIp socket wrappers, config/TLS decl), MemIf,
  SecOC config + Lcfg rewrite, NvM ECC API, SomeIp*, CanNm/LinNm,
  CanTp/FrTp, CanTrcv config rewrite, Icu, Fee/Fls/Flash, EthSM/EthIf,
  DoIP/SoAd, Xcp config types, StbM, RamSafety/Platform, RTE + ASW
  (Rte_Read/Write port API, Rte_GetTime, misc)

## Tests
- cmake --build build -j4: 0 errors
- make -C build -j1 -k: 0 errors
- pytest tests/unit: 6/6 passed
- tests/integration fail only on missing .yuleosh/evidence-bundle artifacts
  (CI pipeline outputs; unrelated to compilation)

## Notes
- TODO markers kept for real-platform stubs (mbedtls link, MCAL interrupt
  control, TcpIp socket backend, FCCU fault hook, Uart time source, etc.)
- Flash register defaults are STM32F4-style; board-specific values must be
  adjusted for S32K312 in production

+++ src/bsw/ecual/canNm/include/CanNm_Cfg.h
@@ -180,12 +180,12 @@ extern "C" {
 /**
  * @brief Number of CAN NM channels configured
  */
-#define CANNM_NUMBER_OF_CHANNELS            (uint8)2U
+#define CANNM_NUMBER_OF_CHANNELS            (2U)
 
 /**
  * @brief Maximum number of channels (should match CANNM_NUMBER_OF_CHANNELS)
  */
-#define CANNM_MAX_NUMBER_OF_CHANNELS        (uint8)2U
+#define CANNM_MAX_NUMBER_OF_CHANNELS        (2U)
 
 /*==================================================================================================
  *                                    CHANNEL CONFIGURATION
@@ -287,7 +287,7 @@ extern "C" {
  * Position 1: Second byte
  * Position 0xFF: Node ID not used
  */
-#define CANNM_PDU_NID_POSITION              (uint8)0U
+#define CANNM_PDU_NID_POSITION              (0U)
 
 /**
  * @brief Control Bit Vector (CBV) position in NM PDU
@@ -295,7 +295,7 @@ extern "C" {
  * Position 1: Second byte
  * Position 0xFF: CBV not used
  */
-#define CANNM_PDU_CBV_POSITION              (uint8)1U
+#define CANNM_PDU_CBV_POSITION              (1U)
 
 /**
  * @brief User data start position
